Working batcher added

+++ MSDF Font Library/FontAtlas/JsonBounds.cs
@@ -2,9 +2,9 @@
 {
     public class JsonBounds
     {
-        public float Left { get; set; }
-        public float Bottom { get; set; }
-        public float Right { get; set; }
-        public float Top { get; set; }
+        public double Left { get; set; }
+        public double Bottom { get; set; }
+        public double Right { get; set; }
+        public double Top { get; set; }
     }
 }
