+++ MSDF Font Library/FontAtlas/JsonGlyph.cs
@@ -3,7 +3,7 @@
     public class JsonGlyph
     {
         public int Unicode { get; set; }
-        public float Advance { get; set; }
+        public double Advance { get; set; }
         public JsonBounds PlaneBounds { get; set; }
         public JsonBounds AtlasBounds { get; set; }
     }
